dags/requirements.txt
apache-airflow-providers-mysql==1.1.0